More semantic handling for tag buffered code
For `p= hello`:

Before, the Code is a separate property of the Tag as Tag.code. Now, the
Code is a part of the tag.block. This way, the handlings for that and

p
  = hello

are the same.

When there is a block, e.g.

p= hello
  strong strongly worded hello

before the strong Tag is in the block of the Code, which makes no sense.
Now it is in the Block of the <p> Tag.